backend/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
python-multipart>=0.0.9,<1.0
pandas>=2.2,<3.0
openpyxl>=3.1,<4.0
pypdf>=4.2,<6.0
groq>=0.11,<1.0
python-dotenv>=1.0,<2.0
weasyprint>=62.3,<70.0
